windows fix try

+++ pennylane_lightning/core/src/simulators/lightning_qubit/gates/OpToMemberFuncPtr.hpp
@@ -43,7 +43,7 @@ template <class PrecisionT, class ParamT, class GateImplementation,
           GateOperation gate_op>
 struct GateOpToMemberFuncPtr {
     // raises compile error when this struct is instantiated.
-    static_assert(sizeof(PrecisionT) == std::numeric_limits<std::size_t>::max(),
+    static_assert(static_cast<std::size_t>(sizeof(PrecisionT)) == static_cast<std::size_t>(std::numeric_limits<std::size_t>::max()),
                   "GateOpToMemberFuncPtr is not defined for the given gate. "
                   "When you define a new GateOperation, check that you also "
                   "have added the corresponding entry in "
